pull

+++ fast_food/database/migrations/2022_05_01_144058_create_chi_tiet_don_hangs_table.php
@@ -15,7 +15,7 @@ class CreateChiTietDonHangsTable extends Migration
     {
         Schema::create('chi_tiet_don_hangs', function (Blueprint $table) {
             $table->id();
-            $table->decimal('don_gia', 9, 3);
+            $table->double('don_gia');
             $table->integer('so_luong');
             $table->double('thanh_tien');
             $table->unsignedBigInteger('don_hang_id')->nullable();//khoá ngoại
